+++ GameJam_SFML_022023/plant.h
@@ -5,15 +5,15 @@ class Game;
 class RootSegment : public sf::RectangleShape
 {
 public:
-	RootSegment(sf::Vector2f position, double size, double speed, const sf::Texture *texture);
+	RootSegment(sf::Vector2f position, float size, float speed, const sf::Texture *texture);
 	void			update(float dt);
 
 	bool			 get_active() const { return _active; }
 	void			set_active(bool value) { _active = value; }
 private:
 	bool			_active;
-	double			_target_y_size;
-	double			_texture_y_scale;
+	float			_target_y_size;
+	float			_texture_y_scale;
 	sf::Vector2f	_position;
 	sf::Vector2f	_speed;
 };
@@ -21,23 +21,23 @@ private:
 class Root
 {
 public:
-	Root(sf::Vector2f position, double size, double speed, const Game *game);
+	Root(sf::Vector2f position, float size, float speed, const Game *game);
 	void			update(float dt);
 	void			draw(sf::RenderWindow& window);
 	void			translate(sf::Vector2f offset);
 	float get_pos();
 	sf::FloatRect	getGlobalBounds();
 	int intersects(sf::FloatRect collider);
-	void			cut(double height);
+	void			cut(float height);
 	bool			getState();
 	void			cut(int i);
 private:
-	double			_speed;
+	float			_speed;
 	bool			_killed;
 	bool			_grounded;
 	sf::RectangleShape _base;
 	std::vector<RootSegment> _segments;
-	double			_alive;
+	float			_alive;
 	const sf::Texture *_tex;
 	sf::RectangleShape _root_base;
 };
